requirements.txt
absl-py==2.2.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.89
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
emoji==1.4.2
etils==1.5.2
eval_type_backport==0.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
flax==0.8.5
fonttools==4.58.1
fsspec==2024.10.0
ftfy==6.0.3
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.26.5
humanize==4.12.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipython==8.12.3
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
meson==1.6.0
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
normalizer==0.0.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.6.4
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.12.3
sympy==1.13.1
tensorstore==0.1.69
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.151
ultralytics-thop==2.0.14
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.21.0
